Harden generated-code portability and 32-bit offset validation
- validate list start offsets in the unsigned domain: on 32-bit targets
  int(uint32) wraps offsets past 2^31 negative, slipping malformed input
  through the validity checks as an empty list (buffer + stream paths)
- clamp reflection list/bitlist read caps to the platform int range
  instead of rejecting large limits; divide bitlist byte caps in uint64
  before capping
- type limit literals past the portable int range as uint64 in error
  constructor arguments so the output compiles on all targets
- replace bare min/max builtins and int(min(...)) capacity clamps in
  generated code with sszutils.Min/Max/CapToInt; a target package can
  shadow the builtins with its own functions
- drop max(x, 0) around provably non-negative counts
- ride int(ctx.exprs[N]) conversions in the streaming encoder with
  analyzer suppressions (#nosec G115 + lgtm), emitted at the build sites

+++ codegen/parser.go
@@ -1451,7 +1451,7 @@ func (p *Parser) buildVectorDescriptor(desc *ssztypes.TypeDescriptor, dataType, 
 	switch t := schemaType.(type) {
 	case *types.Array:
 		schemaElemType = t.Elem()
-		length = int64(t.Len())
+		length = t.Len()
 		if len(sizeHints) > 0 && sizeHints[0].Size > 0 {
 			byteSize := sizeHints[0].Size
 			if sizeHints[0].Bits {
